+++ test/sha1_test.dart
@@ -8,22 +8,6 @@ import "package:crypto/crypto.dart";
 import "package:test/test.dart";
 
 void main() {
-  group("with the old API", () {
-    test('add may not be called after close', () {
-      var sha = new SHA1();
-      sha.close();
-      expect(() => sha.add([0]), throwsStateError);
-    });
-
-    test('close returns the same digest repeatedly', () {
-      var sha = new SHA1();
-      var digest = sha.close();
-      expect(sha.close(), equals(digest));
-      expect(sha.close(), equals(digest));
-      expect(sha.close(), equals(digest));
-    });
-  });
-
   group("with a chunked converter", () {
     test('add may not be called after close', () {
       var sink = sha1.startChunkedConversion(new StreamController().sink);